refactor(format): split init.lua into submodules
